+++ test/tools/yulInterpreter/EVMInstructionInterpreter.cpp
@@ -42,7 +42,7 @@ u256 exp256(u256 _base, u256 _exponent)
 	u256 result = 1;
 	while (_exponent)
 	{
-		if (static_cast<limb_type>(_exponent) & 1)	// If exponent is odd.
+		if (boost::multiprecision::bit_test(_exponent, 0))
 			result *= _base;
 		_base *= _base;
 		_exponent >>= 1;
@@ -441,6 +441,29 @@ u256 EVMInstructionInterpreter::eval(
 	}
 	}
 
+	return 0;
+}
+
+u256 EVMInstructionInterpreter::evalBuiltin(YulString _name, const std::vector<u256>& _arguments)
+{
+	if (_name == "datasize"_yulstring)
+		return u256(keccak256(h256(_arguments.at(0)))) & 0xfff;
+	else if (_name == "dataoffset"_yulstring)
+		return u256(keccak256(h256(_arguments.at(0) + 2))) & 0xfff;
+	else if (_name == "datacopy"_yulstring)
+	{
+		// This is identical to codecopy.
+		if (logMemoryWrite(_arguments.at(0), _arguments.at(2)))
+			copyZeroExtended(
+				m_state.memory,
+				m_state.code,
+				size_t(_arguments.at(0)),
+				size_t(_arguments.at(1) & size_t(-1)),
+				size_t(_arguments.at(2))
+			);
+	}
+	else
+		yulAssert(false, "Unknown builtin: " + _name.str());
 	return 0;
 }